SDK-293: Refactor SDK so Tomcat is launched in a separate process
