+++ maven-plugin/src/main/java/org/openmrs/maven/plugins/Info.java
@@ -3,7 +3,7 @@ package org.openmrs.maven.plugins;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.openmrs.maven.plugins.model.Server;
-import org.openmrs.maven.plugins.utility.Project;
+import org.openmrs.maven.plugins.model.Project;
 
 import java.util.Set;
