+++ maven-plugin/src/main/java/org/openmrs/maven/plugins/utility/SpaInstaller.java
@@ -2,10 +2,7 @@ package org.openmrs.maven.plugins.utility;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.apache.commons.lang.StringUtils;
-import org.apache.maven.execution.MavenSession;
-import org.apache.maven.plugin.BuildPluginManager;
 import org.apache.maven.plugin.MojoExecutionException;
-import org.apache.maven.project.MavenProject;
 import org.openmrs.maven.plugins.model.DistroProperties;
 
 import java.io.File;
