--- maven-plugin/src/main/resources/sdk-stats.properties
@@ -1,22 +1,0 @@
-statsEnabled=false
-statsLastReported=
-statsUser=
-statsLastUsed=
-statsSetupCalls=0
-statsRunCalls=0
-statsBuildCalls=0
-statsCreate-projectCalls=0
-statsDeleteCalls=0
-statsDeployCalls=0
-statsHelpCalls=0
-statsInfoCalls=0
-statsPrCalls=0
-statsReleaseCalls=0
-statsCloneCalls=0
-statsPullCalls=0
-statsResetCalls=0
-statsUndeployCalls=0
-statsWatchCalls=0
-statsUnwatchCalls=0
-statsSdkVersion=
-statsBuild-distroCalls=0
