--- maven-plugin/src/main/resources/sdk.properties
@@ -1,3 +1,0 @@
-groupId=${groupId}
-artifactId=${artifactId}
-version=${version}
