+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/model/Artifact.java
