+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/model/BaseSdkProperties.java
