+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/model/Project.java
@@ -1,4 +1,4 @@
-package org.openmrs.maven.plugins.utility;
+package org.openmrs.maven.plugins.model;
 
 import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.builder.EqualsBuilder;
@@ -8,7 +8,6 @@ import org.apache.maven.model.Parent;
 import org.apache.maven.model.io.xpp3.MavenXpp3Reader;
 import org.apache.maven.plugin.MojoExecutionException;
 import org.codehaus.plexus.util.xml.pull.XmlPullParserException;
-import org.openmrs.maven.plugins.model.Artifact;
 
 import java.io.File;
 import java.io.FileReader;
