+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/model/SdkStatistics.java
