+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/utility/DistroHelper.java
@@ -1,7 +1,6 @@
 package org.openmrs.maven.plugins.utility;
 
 import org.apache.commons.io.FileUtils;
-import org.apache.commons.io.IOUtils;
 import org.apache.commons.lang.StringUtils;
 import org.apache.maven.execution.MavenSession;
 import org.apache.maven.plugin.BuildPluginManager;
