+++ sdk-commons/src/main/java/org/openmrs/maven/plugins/utility/Wizard.java
@@ -1,7 +1,5 @@
 package org.openmrs.maven.plugins.utility;
 
-import javax.annotation.Nullable;
-
 import org.apache.maven.plugin.MojoExecutionException;
 import org.openmrs.maven.plugins.model.DistroProperties;
 import org.openmrs.maven.plugins.model.Server;
@@ -27,11 +25,11 @@ public interface Wizard {
 
     void promptForRefAppVersionIfMissing(Server server, VersionsHelper versionsHelper) throws MojoExecutionException;
 
-    void promptForRefAppVersionIfMissing(Server server, VersionsHelper versionsHelper, @Nullable String customMessage) throws MojoExecutionException;
+    void promptForRefAppVersionIfMissing(Server server, VersionsHelper versionsHelper, String customMessage) throws MojoExecutionException;
 
     String promptForRefAppVersion(VersionsHelper versionsHelper) throws MojoExecutionException;
 
-    String promptForDistroVersion(String distroGroupId, String distroArtifactId, String distroVersion, String distroName, VersionsHelper versionsHelper, @Nullable String customMessage)
+    String promptForDistroVersion(String distroGroupId, String distroArtifactId, String distroVersion, String distroName, VersionsHelper versionsHelper, String customMessage)
             throws MojoExecutionException;
 
     String promptForDistroVersion(String distroGroupId, String distroArtifactId, String distroVersion, String distroName, VersionsHelper versionsHelper)
