+++ sdk-commons/src/main/resources/sdk-stats.properties
