+++ sdk-commons/src/main/resources/sdk.properties
@@ -1,0 +1,3 @@
+groupId=${sdk.groupId}
+artifactId=${sdk.artifactId}
+version=${version}
