+++ sdk-commons/src/test/java/org/openmrs/maven/plugins/model/SdkStatisticsTest.java
