+++ sdk-commons/src/test/java/org/openmrs/maven/plugins/utility/DefaultWizardTest.java
