+++ tomcat-maven-plugin/src/main/java/org/openmrs/maven/plugins/RunTomcat.java
@@ -20,11 +20,11 @@ import org.apache.maven.project.MavenProject;
 import org.codehaus.plexus.classworlds.ClassWorld;
 import org.codehaus.plexus.classworlds.realm.ClassRealm;
 import org.codehaus.plexus.classworlds.realm.DuplicateRealmException;
+import org.openmrs.maven.plugins.model.Project;
 import org.openmrs.maven.plugins.model.Server;
 import org.openmrs.maven.plugins.model.Version;
 import org.openmrs.maven.plugins.utility.DockerHelper;
 import org.openmrs.maven.plugins.utility.OpenMRSSDKRedirectServlet;
-import org.openmrs.maven.plugins.utility.Project;
 import org.openmrs.maven.plugins.utility.SDKConstants;
 import org.openmrs.maven.plugins.utility.Wizard;
 
@@ -148,6 +148,7 @@ public class RunTomcat extends AbstractMojo {
 		if (port == null) {
 			port = 8080;
 		}
+		
 		Context root = tomcat.addContext("", tempDirectory.getAbsolutePath());
 		tomcat.addServlet("", "openmrsRedirectServlet", new OpenMRSSDKRedirectServlet());
 		root.addServletMappingDecoded("/*", "openmrsRedirectServlet");
